huggingface/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langchain-google-genai==2.1.12
langchain-huggingface
sentence-transformers
torch>=2.6.0
faiss-cpu==1.12.0
pypdf==4.0.0
pydantic==2.9.2